Updated message checker algorithm
In complex message parser, non alphabetical chars are not removed. instead I send back the original msg.
In MessageChecker, I split up the original msg by non alphanumerics, then I check if each word is a legitimate word.
If each word are legitimate, then it is not used in a bad context.
If there are words that are not legitimate, then concat the word and test one more time.