Fixed timeout reporting for unban/unmutes
The bot would report the duration of ban/mute based on the duration passed into the setTimeout function, which is wrong if the bot restarts and a shortened duration is set.

This commit fixes that by adding the actual start time to all relevant methods and the timeout table in the database.